pyproject.toml
[tool.poetry]
name = "rag-4-ird"
version = "0.1.0"
description = ""
authors = ["Savith Panamgama <geethikasavith@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.54.3"
chromadb = "^0.5.18"
langchain-openai = "^0.2.6"
langchain-text-splitters = "^0.3.2"
python-dotenv = "^1.0.1"
langgraph = "^0.2.45"
langchain = "^0.3.7"
uvicorn = "^0.32.0"
fastapi = "^0.115.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
